# docs/swagger.yaml
definitions:
...
  models.BusinessCategory:
    enum:
    - B2B
    - B2C
    - Non_Profit
    - Government
    - Mixed
    type: string
    x-enum-varnames:
    - BusinessCategoryB2B
    - BusinessCategoryB2C
    - BusinessCategoryNonProfit
    - BusinessCategoryGovernment
    - BusinessCategoryMixed
...
  models.BusinessPhase:
    enum:
    - Startup
    - Growth
    - Mature
    - Exit
    type: string
    x-enum-varnames:
    - BusinessPhaseStartup
    - BusinessPhaseGrowth
    - BusinessPhaseMature
    - BusinessPhaseExit
...
  models.BusinessType:
    enum:
    - Consulting
    - Retail
    - Technology
    - Manufacturing
    - Services
    - Other
    type: string
    x-enum-varnames:
    - BusinessTypeConsulting
    - BusinessTypeRetail
    - BusinessTypeTechnology
    - BusinessTypeManufacturing
    - BusinessTypeServices
    - BusinessTypeOther
...
  models.ProjectMemberRole:
    enum:
    - manager
    - contributor
    - reviewer
    type: string
    x-enum-varnames:
    - ProjectMemberRoleManager
    - ProjectMemberRoleContributor
    - ProjectMemberRoleReviewer
...
  models.ProjectStatus:
    enum:
    - planning
    - active
    - on_hold
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - ProjectStatusPlanning
    - ProjectStatusActive
    - ProjectStatusOnHold
    - ProjectStatusCompleted
    - ProjectStatusCancelled
...
  ports.AddProjectMemberInput:
    properties:
      project_id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/models.ProjectMemberRole'
        enum:
        - manager
        - contributor
        - reviewer
      user_id:
        type: integer
    required:
    - project_id
    - role
    - user_id
    type: object
...
  ports.BusinessResponse:
    properties:
      active:
        type: boolean
      address:
        type: string
      business_category:
        $ref: '#/definitions/models.BusinessCategory'
      business_phase:
        $ref: '#/definitions/models.BusinessPhase'
      business_type:
        $ref: '#/definitions/models.BusinessType'
      city:
        type: string
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone_no:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      operator_user:
        $ref: '#/definitions/ports.UserResponse'
      operator_user_id:
        type: integer
      postal_code:
        type: string
      state:
        type: string
      tagline:
        type: string
      updated_at:
        type: string
      value:
        type: number
      website:
        type: string
    type: object
...
  ports.ProjectMemberResponse:
    properties:
      joined_at:
        type: string
      project:
        $ref: '#/definitions/ports.ProjectResponse'
      project_id:
        type: integer
      role:
        $ref: '#/definitions/models.ProjectMemberRole'
      user:
        $ref: '#/definitions/ports.UserResponse'
      user_id:
        type: integer
    type: object
  ports.ProjectMembersResponse:
    properties:
      count:
        type: integer
      members:
        items:
          $ref: '#/definitions/ports.ProjectMemberResponse'
        type: array
    type: object
...
  ports.ProjectResponse:
    properties:
      actual_end_date:
        type: string
      business:
        $ref: '#/definitions/ports.BusinessResponse'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      manager:
        $ref: '#/definitions/ports.UserResponse'
      members:
        items:
          $ref: '#/definitions/ports.ProjectMemberResponse'
        type: array
      name:
        type: string
      project_status:
        $ref: '#/definitions/models.ProjectStatus'
      regions:
        items:
          $ref: '#/definitions/ports.RegionResponse'
        type: array
      start_date:
        type: string
      target_end_date:
        type: string
      updated_at:
        type: string
    type: object
...
  ports.RegionResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  ports.UpdateProjectMemberRoleInput:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/models.ProjectMemberRole'
        enum:
        - manager
        - contributor
        - reviewer
    required:
    - role
    type: object
...
  ports.UserResponse:
    properties:
      active:
        type: boolean
      contact_email:
        type: string
      contact_phone_no:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      login_email:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /projects/{id}/members:
    get:
      description: Retrieves a list of all members associated with a project. Accessible
        by any authenticated user.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of project members
          schema:
            $ref: '#/definitions/ports.ProjectMembersResponse'
        "400":
          description: Invalid project ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Project Members
      tags:
      - projects
      - members
    post:
      consumes:
      - application/json
      description: Adds a user to a project with a specified role. Only accessible
        by the **Project Manager**.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Member details (UserID, Role)
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/ports.AddProjectMemberInput'
      produces:
      - application/json
      responses:
        "201":
          description: Member added successfully
          schema:
            $ref: '#/definitions/ports.ProjectMemberResponse'
        "400":
          description: Invalid project ID, request body, or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the project manager)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectNotFound or ErrUserNotFound
          schema:
            additionalProperties: true
            type: object
        "409":
          description: ErrProjectMemberAlreadyExists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add Project Member
      tags:
      - projects
      - members
  /projects/{id}/members/{userID}:
    delete:
      description: Removes a user from a project. Allowed for the **Project Manager**
        (to remove anyone) or the **User** (to remove self).
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID of the member to remove
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Member removed successfully (No Content)
        "400":
          description: Invalid ID or ErrCannotRemoveManager
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the manager and not the user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectMemberNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove Project Member
      tags:
      - projects
      - members
    get:
      description: Retrieves a specific project member record by Project ID and User
        ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID of the member
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Member retrieved successfully
          schema:
            $ref: '#/definitions/ports.ProjectMemberResponse'
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectMemberNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Specific Project Member
      tags:
      - projects
      - members
    put:
      consumes:
      - application/json
      description: Updates the role of an existing project member. Only accessible
        by the **Project Manager**.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID of the member
        in: path
        name: userID
        required: true
        type: integer
      - description: New role for the member
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/ports.UpdateProjectMemberRoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: Member role updated successfully
          schema:
            $ref: '#/definitions/ports.ProjectMemberResponse'
        "400":
          description: Invalid ID, request body, or ErrInvalidRole
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the project manager)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrProjectMemberNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Project Member Role
      tags:
      - projects
      - members
...
  /users/{id}/project-memberships:
    get:
      description: Retrieves a list of all projects the specified user is a member
        of. Requires self-management.
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by member role (manager, contributor, reviewer)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of project memberships
          schema:
            $ref: '#/definitions/ports.ProjectMembersResponse'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Projects by User
      tags:
      - users
      - members
...
swagger: "2.0"